Fix all source code, Update in 2018-4-5
